requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
aws-cdk-lib==2.81.0
aws-cdk.asset-awscli-v1==2.2.184
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.155
boto3==1.28.9
botocore==1.31.9
cattrs==22.2.0
certifi==2023.5.7
charset-normalizer==3.2.0
constructs==10.2.34
frozenlist==1.4.0
idna==3.4
importlib-resources==5.12.0
jmespath==1.0.1
jsii==1.82.0
multidict==6.0.4
mysql-connector-python==8.0.32
openai==0.27.8
protobuf==3.20.3
publication==0.0.3
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
tqdm==4.65.0
typeguard==2.13.3
# typing_extensions @ file:///private/tmp/python-typing-extensions-20230214-6416-1eblvgh/typing_extensions-4.5.0/dist/typing_extensions-4.5.0-py3-none-any.whl
urllib3==1.26.16
yarl==1.9.2
